pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=1.0.0",
    "langchain-openai>=0.3.30",
    "python-dotenv>=1.1.1",
    "langchain-classic>=1.0.1",
]
